+++ inst/framework7-5.3.0/input-bindings/gaugeInputBinding.js
@@ -37,6 +37,9 @@ $.extend(f7GaugeBinding, {
       }
     });
 
+    // valueText should not be let as a free parameter since it does not make sense
+    // that the value is different from the displayed text.
+    data.valueText = 100 * data.value + '%';
     // add the id
     data.el = '#' + id;
 
@@ -58,7 +61,19 @@ $.extend(f7GaugeBinding, {
     var g = app.gauge.get($(el));
     g.update({
       value: data.value / 100,
-      valueText: data.value + '%'
+      valueText: data.value + '%',
+      labelText: data.text,
+      size: data.size,
+      bgColor: data.bgColor,
+      borderBgColor: data.borderBgColor,
+      borderColor: data.borderColor,
+      borderWidth: data.borderWidth,
+      valueTextColor: data.valueTextColor,
+      valueFontSize: data.valueFontSize,
+      valueFontWeight: data.valueFontWeight,
+      labelTextColor: data.labelTextColor,
+      labelFontSize: data.labelFontSize,
+      labelFontWeight: data.labelFontWeight
     });
   }
 });